Parse kaktovik slash notation
This commit adds the ability to parse numbers from the base 20
kaktovik inupiaq numeral system written in the slash notation
described at the top of `src/kaktovik_inupiaq.c`.

Originally the commit was going to focus on adding floating point
parsing for binary numbers but that ended up not happening. So
`bin.c` is a little messy at the moment.

For the kaktovik testing function I added an extra helper function
to remove the repetition I was getting with the macro test
approach. Though, I ended up only using `parse_test64`(the
function in question) for the kaktovik testing. In later commits
I'll transition to using it in the other tests as well.